requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==2.14.6
dill==0.3.7
filelock==3.20.3
Flask==2.3.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2023.10.0
huggingface_hub==0.36.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.1
multiprocess==0.70.15
nltk==3.9.2
numpy==1.24.3
opencv-python==4.8.1.78
packaging==26.0
pandas==2.0.3
Pillow==10.0.1
propcache==0.4.1
pyarrow==12.0.1
pyparsing==3.3.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
